@@ -1,6 +1,6 @@
 {
     "name": "boltz-frontend",
-    "version": "1.8.7",
+    "version": "1.8.8",
     "description": "Boltz Exchange, Atomic Swaps, Web App",
     "scripts": {
         "start": "npm run regtest && vite",
@@ -33,29 +33,28 @@
         "check:assets": "node check-assets.mjs"
     },
     "devDependencies": {
-        "@eslint/js": "^9.37.0",
-        "@playwright/test": "^1.56.0",
+        "@eslint/js": "^9.39.1",
+        "@playwright/test": "^1.56.1",
         "@solidjs/testing-library": "^0.8.10",
         "@testing-library/jest-dom": "^6.9.1",
         "@testing-library/user-event": "^14.6.1",
-        "@trivago/prettier-plugin-sort-imports": "^5.2.2",
-        "@types/node": "^24.7.0",
+        "@trivago/prettier-plugin-sort-imports": "^6.0.0",
+        "@types/node": "^24.10.1",
         "@webbtc/webln-types": "^3.0.0",
-        "axios": "^1.12.2",
-        "eslint": "^9.37.0",
+        "axios": "^1.13.2",
+        "eslint": "^9.39.1",
         "eslint-plugin-solid": "^0.14.5",
         "git-cliff": "^2.10.1",
-        "globals": "^16.4.0",
-        "jsdom": "^27.1.0",
+        "globals": "^16.5.0",
         "prettier": "^3.6.2",
-        "sass-embedded": "^1.93.2",
+        "sass-embedded": "^1.93.3",
         "svgo": "^4.0.0",
         "typescript": "^5.9.3",
-        "typescript-eslint": "^8.46.0",
-        "vite": "^7.1.9",
+        "typescript-eslint": "^8.46.4",
+        "vite": "^7.2.2",
         "vite-plugin-mkcert": "^1.17.9",
         "vite-plugin-node-polyfills": "^0.24.0",
-        "vite-plugin-solid": "^2.11.9",
+        "vite-plugin-solid": "^2.11.10",
         "vite-plugin-top-level-await": "^1.6.0",
         "vite-plugin-wasm": "^3.5.0",
         "vitepress": "^1.6.4",
@@ -65,17 +64,17 @@
         "@bitcoinerlab/secp256k1": "^1.2.0",
         "@fontsource/noto-mono": "^5.2.5",
         "@fontsource/noto-sans": "^5.2.10",
-        "@ledgerhq/hw-app-eth": "^6.46.0",
-        "@ledgerhq/hw-transport-webhid": "^6.30.8",
+        "@ledgerhq/hw-app-eth": "^6.47.0",
+        "@ledgerhq/hw-transport-webhid": "^6.30.9",
         "@reown/appkit": "1.6.9",
         "@reown/appkit-adapter-ethers": "1.6.9",
-        "@scure/base": "^1.2.6",
-        "@scure/bip32": "^1.7.0",
-        "@scure/bip39": "^1.6.0",
+        "@scure/base": "^2.0.0",
+        "@scure/bip32": "^2.0.1",
+        "@scure/bip39": "^2.0.1",
         "@solid-primitives/i18n": "^2.2.1",
         "@solid-primitives/storage": "^4.3.3",
-        "@solidjs/router": "^0.15.3",
-        "@trezor/connect-web": "^9.6.3",
+        "@solidjs/router": "^0.15.4",
+        "@trezor/connect-web": "^9.6.4",
         "@vulpemventures/secp256k1-zkp": "^3.2.1",
         "bignumber.js": "^9.3.1",
         "bitcoinjs-lib": "^6.1.7",
@@ -93,6 +92,6 @@
         "qr-scanner": "^1.4.2",
         "qrcode": "^1.5.4",
         "solid-icons": "^1.1.0",
-        "solid-js": "1.9.9"
+        "solid-js": "1.9.10"
     }
 }
