@@ -4,7 +4,7 @@ import { Show, createMemo, createResource } from "solid-js";
 import CopyButton from "../components/CopyButton";
 import QrCode from "../components/QrCode";
 import { BTC } from "../consts/Assets";
-import type { SwapType } from "../consts/Enums";
+import { type SwapType } from "../consts/Enums";
 import { useGlobalContext } from "../context/Global";
 import { getPairs } from "../utils/boltzClient";
 import { formatAmount, formatDenomination } from "../utils/denomination";
