@@ -4,7 +4,7 @@ import {
     mnemonicToSeedSync,
     validateMnemonic,
 } from "@scure/bip39";
-import { wordlist } from "@scure/bip39/wordlists/english";
+import { wordlist } from "@scure/bip39/wordlists/english.js";
 
 export enum Errors {
     InvalidFile = "invalid file",
